% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparetree.R
\name{comparetree}
\alias{comparetree}
\title{comparetree}
\usage{
comparetree(tree, id1 = 2L, id2 = 3L, type = "detect",
  highlights = 5, plot = FALSE, layout = "constrained", ...)
}
\arguments{
\item{tree}{a networktree object}

\item{id1}{the first partition}

\item{id2}{the second partition}

\item{type}{"cor", "pcor", or "glasso". Defaults to automatic detection}

\item{highlights}{the number of comparisons to highlight}

\item{plot}{plot a comparison of the two partitions?}

\item{layout}{layout for the plots. The default "constrained" uses a
FR layout from the full dataset}

\item{...}{additional arguments passed to qgraph}
}
\description{
Quickly compares two partitions of a networktree object
}
\examples{

}
