% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cortrafo.R
\name{cortrafo}
\alias{cortrafo}
\title{cortrafo}
\usage{
cortrafo(data, weights, control, model, ...)
}
\arguments{
\item{data}{a matrix or data}

\item{weights}{not currently used}

\item{control}{not currently used}

\item{model}{can be any combination of c("correlation", "mean", "variance"). 
Scores are determined based on the specified characteristics}

\item{...}{not currently used}
}
\description{
An influence function that transforms the response variables (y1, y2, y3...) into scores 
relevant to the correlations, means, and/or variances. 

For example, in the case of correlations the variables are transformed into a matrix of 
(n^2-n)/2 columns (e.g., the number of total correlations), and i rows, where i is the # of 
observations of y1, where the mean of each vector is equal to the correlation between y1 and y2, y1 and y3, etc.

Used internally in when method="ctree".
}
