% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctreeNetwork.R
\name{ctreeNetwork}
\alias{ctreeNetwork}
\title{Conditional inference trees for networks}
\usage{
ctreeNetwork(formula, data, weights = NULL, type = c("cor", "pcor",
  "EBICglasso"), na.action, ...)
}
\arguments{
\item{formula}{A symbolic description of the model to be fit. This
should either be of type \code{y1 + y2 + y3 ~ x1 + x2} with reponse
vectors \code{y1}, \code{y2}, and \code{y3} or \code{y ~ x1 + x2}
with a matrix response {y}. \code{x1} and \code{x2} are used as
partitioning variables.}

\item{data}{a data frame containing the variables in the model}

\item{weights}{weights}

\item{type}{the type of network to compute. Can be "cor", "pcor", or "EBICglasso"}

\item{na.action}{a function which indicates what should happen when the data
contain missing values (\code{NA}s).}

\item{...}{arguments passed to \code{\link[partykit]{mob_control}}}
}
\value{
\code{\link{ctreeNetwork}} returns a list of class "\code{}" which contains:
}
\description{
This function is equivalent to networktreeConditional, but using the
traditional partykit interface.
}
\details{
Computes a tree model with networks as the end branches using conditional inference.
Wraps the ctree function from partykit.

A high-level convenience interface to ctree() + mvnfit()
}
