% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnfit.R
\name{mvnfit}
\alias{mvnfit}
\title{Maximum Likelihood Estimation for Multivariate Normal Model}
\usage{
mvnfit(y, x = NULL, start = NULL, weights = NULL, offset = NULL,
  cor = FALSE, ..., estfun = FALSE, object = FALSE)
}
\arguments{
\item{y}{A matrix or data.frame where each row corresponds to a k-dim observation.}

\item{x}{Not used yet}

\item{start}{Not used yet}

\item{weights}{Not used yet}

\item{offset}{Not used yet}

\item{cor}{Logical. Should only be the parameter estimates and scores pertaining
to the correlations be returned (i.e., those for mu and sigma omitted)?}

\item{...}{Not used yet}

\item{estfun}{Logical. Should the matrix of score contributions (aka estimating
functions) be returned?}

\item{object}{Not used yet}
}
\description{
Fit a multivariate normal model without covariates or covariance restrictions.
In addition to the (straightforward) parameter estimates the fitted log-likelihood
and corresponding score contributions are computed.
}
\details{
Used internally in mobNetwork and networktreeModelBased
}
