% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networktreeConditional.R
\name{networktreeConditional}
\alias{networktreeConditional}
\title{Conditional inference trees for networks}
\usage{
networktreeConditional(data, splitVars, weights = NULL, type = c("cor",
  "pcor", "EBICglasso"), na.action, ...)
}
\arguments{
\item{data}{the cross-sectional dataset from which to compute the network}

\item{splitVars}{the variables with which to test network splitting. Can be vector, matrix, or dataframe}

\item{weights}{weights}

\item{type}{the type of network to compute. Can be "cor", "pcor", or "EBICglasso"}

\item{na.action}{a function which indicates what should happen when the data
contain missing values (\code{NA}s).}

\item{...}{additional arguments passed cortrafo}
}
\value{
\code{\link{networktreeConditional}} returns a list of class "\code{}" which contains:
}
\description{
Computes a tree model with networks as the end branches using conditional inference.
Wraps the ctree function from partykit. This approach is slow; for a faster method
see networkModelBased.
}
\examples{
\donttest{
set.seed(1)
d <- data.frame(trend = 1:200, foo = runif(200, -1, 1))
d <- cbind(d, rbind(
  rmvnorm(100, mean = c(0, 0, 0),
          sigma = matrix(c(1, 0.5, 0.5, 0.5, 1, 0.5, 0.5, 0.5, 1), ncol = 3)),
  rmvnorm(100, mean = c(0, 0, 0),
          sigma = matrix(c(1, 0, 0.5, 0, 1, 0.5, 0.5, 0.5, 1), ncol = 3))
))
colnames(d)[3:5] <- paste0("y", 1:3)

## Now use the function
tree1 <- networktreeConditional(netdata=d[,3:5], splitVars=d[,1:2])

}
}
