% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networktree.R
\name{plot.networktree}
\alias{plot.networktree}
\title{Plotting 'networktree' objects}
\usage{
\method{plot}{networktree}(
  x,
  type = "detect",
  transform = NULL,
  layout = "lock",
  partyargs = list(),
  sdbars = NULL,
  customplotfunction = NULL,
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{x}{an object of type 'networktree'}

\item{type}{the type of plot. The currently available options are "network" and 
"barplot" (for means and/or variances). By default, "network" is used unless
the correlations were not used to split in networktree.}

\item{transform}{"cor", "pcor", or "glasso". If set to NULL, transform detected from x}

\item{layout}{network layout, passed to qgraph. Default "lock" computes spring 
layout for the full sample and applies this to all graphs}

\item{partyargs}{additional arguments (list format) passed to \code{partykit::plot.party}}

\item{sdbars}{if type="barplot", should std deviation error bars be plotted?}

\item{customplotfunction}{to create custom plots in the terminal nodes, provide a base R 
plotting function that takes partitioned data as input}

\item{na.rm}{should NA values be removed prior to calculating relevant parameters?}

\item{...}{additional arguments passed to qgraph or barplot}
}
\description{
Wraps plot.party to plot a tree model with networks on the ends. Networks
are plotted with qgraph, and additional arguments are passed there
}
