% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networktree.R
\name{predict.networktree}
\alias{predict.networktree}
\title{Predict 'networktree' objects}
\usage{
\method{predict}{networktree}(object, newdata = NULL, type = c("node", "parameter"), ...)
}
\arguments{
\item{object}{a fitted 'networktree'}

\item{newdata}{An optional data frame in which to look for variables with
which to predict.  If omitted, the fitted values are used.}

\item{type}{"node", or "parameter". Specifies whether to predict nodes
(return value is a vector) or parameters (matrix).}

\item{...}{not used}
}
\description{
Wraps predict.party
}
