% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networktree.R
\name{print.networktree}
\alias{print.networktree}
\title{Printing 'treenetwork' objects}
\usage{
\method{print}{networktree}(x, parameters = FALSE, FUN = NULL, ...)
}
\arguments{
\item{x}{an object of type 'networktree'}

\item{parameters}{print parameters for each partition? 
See getnetwork function for extracting parameters conveniently}

\item{FUN}{only evaluated if parameters=TRUE, passed to print.modelparty}

\item{...}{additional arguments passed print.modelparty}
}
\description{
Wraps print.modelparty to print a tree model with networks on the ends.
}
