% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networktree.R
\name{plot.networktree}
\alias{plot.networktree}
\title{Plotting 'treenetwork' objects}
\usage{
\method{plot}{networktree}(x, type = NULL, layout = "circle", ...)
}
\arguments{
\item{x}{an object of type 'networktree'}

\item{type}{"cor", "pcor", or "EBICglasso". If set to NULL, type detected from x}

\item{layout}{argument passed to qgraph. Can be set to custom layout.}

\item{...}{additional arguments passed qgraph}
}
\description{
Wraps plot.party to plot a tree model with networks on the ends. Networks
are plotted with qgraph, and additional arguments are passed there
}
