% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networktree.R
\name{plot.networktree}
\alias{plot.networktree}
\title{Plotting 'networktree' objects}
\usage{
\method{plot}{networktree}(x, transform = NULL, layout = "lock", partyargs = list(), ...)
}
\arguments{
\item{x}{an object of type 'networktree'}

\item{transform}{"cor", "pcor", or "glasso". If set to NULL, transform detected from x}

\item{layout}{network layout, passed to qgraph. Default "lock" computes spring 
layout for the full sample and applies this to all graphs}

\item{partyargs}{additional arguments (list format) passed to \code{partykit::plot.party}}

\item{...}{additional arguments passed to qgraph}
}
\description{
Wraps plot.party to plot a tree model with networks on the ends. Networks
are plotted with qgraph, and additional arguments are passed there
}
